Move wireguard tests to a separate file
